+++ solutions/devsprint-mychelle-rocha-1/.idea/misc.xml
@@ -3,9 +3,11 @@
   <component name="DesignSurface">
     <option name="filePathToZoomLevelMap">
       <map>
+        <entry key="..\:/Users/Adenilson/AndroidStudioProjects/challenge-viewcode-github-android/solutions/devsprint-mychelle-rocha-1/app/src/main/res/drawable/ic_search_toolbar.xml" value="0.159" />
         <entry key="..\:/Users/Adenilson/AndroidStudioProjects/challenge-viewcode-github-android/solutions/devsprint-mychelle-rocha-1/app/src/main/res/layout/activity_main.xml" value="0.20416666666666666" />
         <entry key="..\:/Users/Adenilson/AndroidStudioProjects/challenge-viewcode-github-android/solutions/devsprint-mychelle-rocha-1/app/src/main/res/layout/content_repository_list.xml" value="0.19010416666666666" />
         <entry key="..\:/Users/Adenilson/AndroidStudioProjects/challenge-viewcode-github-android/solutions/devsprint-mychelle-rocha-1/app/src/main/res/layout/repository_header.xml" value="0.19010416666666666" />
+        <entry key="..\:/Users/Adenilson/AndroidStudioProjects/challenge-viewcode-github-android/solutions/devsprint-mychelle-rocha-1/app/src/main/res/layout/repository_list_item.xml" value="0.20416666666666666" />
       </map>
     </option>
   </component>
