+++ solutions/devsprint-mychelle-rocha-1/app/src/main/res/layout/activity_main.xml
@@ -17,8 +17,8 @@
             android:layout_width="match_parent"
             android:layout_height="?attr/actionBarSize"
             android:background="?attr/colorPrimary"
-            app:popupTheme="@style/Theme.GitHubApp.PopupOverlay" />
-
+            app:popupTheme="@style/Theme.GitHubApp.PopupOverlay"
+            app:title="@string/title_toolbar" />
     </com.google.android.material.appbar.AppBarLayout>
 
     <include
