+++ solutions/devsprint-mychelle-rocha-1/app/src/main/res/menu/menu_main.xml
@@ -8,7 +8,7 @@
         android:title="@string/action_settings"
         app:showAsAction="never" />
     <item
-        android:id="@+id/serch_toolbar"
+        android:id="@+id/search_toolbar"
         android:icon="@drawable/ic_search_toolbar"
         app:showAsAction="ifRoom"
         android:title="@string/title_toolbar"
