+++ solutions/devsprint-mychelle-rocha-1/app/src/main/res/values/strings.xml
@@ -8,8 +8,8 @@
     <string name="next">Next</string>
     <string name="previous">Previous</string>
     <string name="hello_second_fragment">Hello second fragment. Arg: %1$s</string>
-    <string name="avatar_owner_repository">Avatar Owner Repository</string>
     <string name="title_toolbar">Repositories</string>
+    <string name="avatar_owner_repository">Avatar Owner Repository</string>
 
     <!--Text About-->
     <string name="title_text_about">About This Repository</string>
